Tech-Recognition/requirements.txt
matplotlib
librosa==0.9.1
tqdm
pandas
numba==0.53.1
numpy==1.19.2
scipy==1.3
PyYAML==5.4.1
tensorboardX
pyloudnorm
setuptools>=41.0.0
g2p_en
resemblyzer
webrtcvad
tensorboard==2.6.0
scikit-learn==0.24.1
scikit-image==0.16.2
textgrid
jiwer
pycwt
PyWavelets
praat-parselmouth==0.3.3
jieba
einops
chardet